Ver.1.6.5 Update.

+++ Controller/NginxcacheclearController.php
@@ -8,7 +8,7 @@
     * @copyright		Copyright 2015, Studio Necomaneki
     * @link			    http://blog.necomaneki.com/ Studio Necomaneki
     * @package			Nginxcacheclear.Controller
-    * @since			  v 1.6.4
+    * @since			  v 1.6.5
     * @license      MIT lincense
     *
     */
@@ -40,7 +40,8 @@ class NginxcacheclearController extends BcPluginAppController {
         // Data Null
         if (!$defcachedir) {
             $this->Nginxcacheclear->read(null, 1);
-            $this->Nginxcacheclear->set('cachedir', '/var/cache/nginx');
+            $this->Nginxcacheclear->set('cachepath','/var/cache/');
+            $this->Nginxcacheclear->set('cachedir', 'nginx');
             $this->Nginxcacheclear->save();
         }
 
@@ -48,7 +49,9 @@ class NginxcacheclearController extends BcPluginAppController {
 
 // Admin index Page Action
     public function admin_index() {
-        $cachedir = $this->Nginxcacheclear->find('all');
+        $cachepath = $this->Nginxcacheclear->find('all', array('fields' => 'cachepath'));
+        $cachedir = $this->Nginxcacheclear->find('all', array('fields' => 'cachedir'));
+        $this->set('cachepath',$cachepath);
         $this->set('cachedir',$cachedir);
         $this->pageTitle = 'Nginxキャッシュクリア管理';
         $this->render('index');
@@ -56,23 +59,23 @@ class NginxcacheclearController extends BcPluginAppController {
 
 // Admin edit page Action
     public function admin_edit() {
-    // not database.
+        // not database.
         if (!$this->data) {
             // Data Search
-            $this->data = $this->Nginxcacheclear->find('first');
+            $this->data  = $this->Nginxcacheclear->find('first');
         } else {
             // Add Data Set
-            $this->Nginxcacheclear->set($this->data);
+            $this->data  = $this->Nginxcacheclear->set($this->data);
             // $id = 1 only update
             if ($this->Nginxcacheclear->id = 1) {
                 // Sanitize
-                $this->data = Sanitize::clean($this->data, array('escape' => false));
+                $this->data = Sanitize::clean($this->data, array('escape' => false, 'odd_spaces', 'encode', 'dollar', 'carriage', 'unicode', 'backslash'));
                 // Add Data Save
                 $this->Nginxcacheclear->save($this->data);
                 $this->Session->setFlash('ディレクトリを更新しました。');
             } else {
-              // Add Data Error
-                $this->Session->setFlash('ディレクトリの更新ができませんでした。');
+            // Add Data Error
+            $this->Session->setFlash('ディレクトリの更新ができませんでした。');
             }
             $this->redirect(array('plugin'=>'nginxcacheclear', 'controller'=>'nginxcacheclear', 'action'=>'index'));
         }
@@ -80,29 +83,53 @@ class NginxcacheclearController extends BcPluginAppController {
         $this->render('edit');
     }
 
+// Admin Nginx Directory Check
+    public function admin_check() {
+        // Model Search - CachePath
+        $ngxcccheck    = $this->Nginxcacheclear->find('first');
+        $ngxpathcheck  = $ngxcccheck['Nginxcacheclear']['cachepath'] . $ngxcccheck['Nginxcacheclear']['cachedir'];
+        if (file_exists($ngxpathcheck)) {
+            $this->setMessage('設定したディレクトリは存在します。');
+        } else {
+            $this->setMessage('設定したディレクトリが存在しません。');
+        }
+        $this->redirect(array('plugin'=>'nginxcacheclear', 'controller'=>'nginxcacheclear', 'action'=>'index'));
+    }
+
 // Admin Nginx Cache Clear Action
     public function admin_clear() {
         App::import('Core', 'Folder');
-        // Model Search - CachePath
-        $ngxcache = $this->Nginxcacheclear->find('first');
-        // CachePath Add
-        $folder = new Folder($ngxcache['Nginxcacheclear']['cachedir'] . DS );
-        $files = $folder->read(true, true, true);
-        // Nginx Cache Directory Search
-        if ($files[preg_match("\x[0-9A-Fa-f]{1,2}")]) {
-            foreach ($files[1] as $file) {
-                @unlink($file);
-            }
-  	    $Folder = new Folder();
-            foreach ($files[0] as $folder) {
-              $Folder->delete($folder);
-            }
-            $this->setMessage('Nginxのキャッシュを削除しました。');
-        } else {
-            $this->setMessage('削除するキャッシュがありませんでした。');
-        }
-        $this->redirect($this->referer());
-    }
 
+        // Nginx Cache Directory
+        $ngxcache  = $this->Nginxcacheclear->find('first');
+        $ngxccpath = $ngxcache['Nginxcacheclear']['cachepath'];
+        $ngxccdir  = $ngxcache['Nginxcacheclear']['cachedir'];
+        $ngxcc     = $ngxccpath . $ngxccdir;
+
+
+        if (!file_exists($ngxcc)) {
+            // Check Directory
+            $this->setMessage('設定したディレクトリが存在しません。');
+        } else {
+             // CachePath Add
+            $folder = new Folder($ngxcc . DS);
+            $files = $folder->read(true, true, true);
+
+            // Nginx Cache Directory Search
+            if ($files[preg_match("\x[0-9A-Fa-f]{1,2}")]) {
+                foreach ($files[1] as $file) {
+                    @unlink($file);
+                }
+                $Folder = new Folder();
+                foreach ($files[0] as $folder) {
+                    $Folder->delete($folder);
+                }
+                $this->setMessage('Nginxのキャッシュを削除しました。');
+            } else {
+                $this->setMessage('削除するキャッシュがありませんでした。');
+            }
+        }
+        $this->redirect(array('plugin'=>'nginxcacheclear', 'controller'=>'nginxcacheclear', 'action'=>'index'));
+    }
 }
 ?>
